... all of them harmless and inconsequential at runtime, but see commit
comments about the error checks in libnyquist that were revealed by the
compiler warnings as never doing anything useful.

+++ lib-src/libmad/frame.c
@@ -144,7 +144,7 @@ int decode_header(struct mad_header *header, struct mad_stream *stream)
   /* layer */
   header->layer = 4 - mad_bit_read(&stream->ptr, 2);
 
-  if (header->layer == 4) {
+  if ((int)header->layer == 4) {
     stream->error = MAD_ERROR_BADLAYER;
     return -1;
   }
